+++ lib/src/utils/event_handler.dart
@@ -37,7 +37,7 @@ void removeEventHandlerListener<A, B>(
   eventHandler.l.removeWhere((g) => identical(f, g));
   if (eventHandler.l.length == lenBefore) {
     // ignore: avoid_print
-    print('[yjs] Tried to remove event handler that doesn\'t exist.');
+    // print('[yjs] Tried to remove event handler that doesn\'t exist.');
   }
 }